requirements.txt
langchain==0.1.0
langchain-community==0.0.1
langchain-core==0.1.0
openai==1.3.0
pypdf==3.17.0
chromadb==0.4.15
sentence-transformers==2.2.2
rouge-score==0.1.2
nltk==3.8.1
streamlit==1.28.0
tqdm==4.66.1
python-dotenv==1.0.0